Add files via upload
We have all come across ATMs in our cities and it is built on Java. This complex project consists offive different classes and is a console-based application. When the system starts the user is prompted with user id and user pin. On entering the details successfully, then ATM functionalities are unlocked. The project allows to perform following operations: